--- odb-compiler/src/util/platform/linux/FileSystem_linux.cpp
@@ -1,24 +1,0 @@
-#include "odbc/util/FileSystem.hpp"
-#include <unistd.h>
-
-namespace odbc {
-
-// ----------------------------------------------------------------------------
-FILE* dupFilePointer(FILE* file)
-{
-    FILE* newFp;
-    int fd = fileno(file);
-    int newFd = dup(fd);
-    if (newFd == -1)
-        goto dup_failed;
-    newFp = fdopen(newFd, "r");
-    if (newFp == nullptr)
-        goto fdopen_failed;
-
-    return newFp;
-
-    fdopen_failed : close(newFd);
-    dup_failed    : return nullptr;
-}
-
-}
